+++ web/datafed-ws.js
@@ -66,7 +66,8 @@ var g_host,
     g_ver_api_major,
     g_ver_api_minor,
     g_ver_api_patch,
-    g_tls;
+    g_tls,
+    g_google_analytics;
 
 const nullfr = Buffer.from([]);
 
@@ -294,7 +295,7 @@ app.get('/ui/welcome', (a_req, a_resp) => {
         const nonce = crypto.randomBytes(16).toString('base64');
         a_resp.locals.nonce = nonce;
         a_resp.setHeader('Content-Security-Policy', `script-src 'nonce-${nonce}'  auth.globus.org`);
-        a_resp.render('index',{nonce:a_resp.locals.nonce, theme:theme,version:g_version,test_mode:g_test});
+        a_resp.render('index',{nonce:a_resp.locals.nonce, theme:theme,version:g_version,test_mode:g_test,...g_google_analytics});
     }
 });
 
@@ -306,7 +307,7 @@ app.get('/ui/main', (a_req, a_resp) => {
         const nonce = crypto.randomBytes(16).toString('base64');
         a_resp.locals.nonce = nonce;
         a_resp.setHeader('Content-Security-Policy', `script-src 'nonce-${nonce}'`);
-        a_resp.render('main',{nonce:a_resp.locals.nonce,user_uid:a_req.session.uid,theme:theme,version:g_version,test_mode:g_test});
+        a_resp.render('main',{nonce:a_resp.locals.nonce,user_uid:a_req.session.uid,theme:theme,version:g_version,test_mode:g_test,...g_google_analytics});
     }else{
         // datafed-user cookie not set, so clear datafed-id before redirect
         //a_resp.clearCookie( 'datafed-id' );
@@ -333,7 +334,7 @@ app.get('/ui/register', (a_req, a_resp) => {
         const nonce = crypto.randomBytes(16).toString('base64');
         a_resp.locals.nonce = nonce;
         a_resp.setHeader('Content-Security-Policy', `script-src 'nonce-${nonce}' auth.globus.org`);
-        a_resp.render('register', {nonce:a_resp.locals.nonce, uid: a_req.session.uid, uname: clean, theme: theme, version: g_version, test_mode: g_test });
+        a_resp.render('register', {nonce:a_resp.locals.nonce, uid: a_req.session.uid, uname: clean, theme: theme, version: g_version, test_mode: g_test, ...g_google_analytics });
     }
 });
 
@@ -368,7 +369,7 @@ app.get('/ui/error', (a_req, a_resp) => {
     const nonce = crypto.randomBytes(16).toString('base64');
     a_resp.locals.nonce = nonce;
     a_resp.setHeader('Content-Security-Policy', `script-src 'nonce-${nonce}'`);
-    a_resp.render('error',{nonce:a_resp.locals.nonce,theme:"light",version:g_version,test_mode:g_test});
+    a_resp.render('error',{nonce:a_resp.locals.nonce,theme:"light",version:g_version,test_mode:g_test,...g_google_analytics});
 });
 
 /* This is the OAuth redirect URL after a user authenticates with Globus
@@ -1823,6 +1824,10 @@ function loadSettings(){
 
         if ( !g_extern_url ){
             g_extern_url = "http"+(g_tls?'s':'')+"://" + g_host + ":" + g_port;
+        }
+
+        if ( config.operations ){
+          g_google_analytics = { enableGoogleAnalytics: config.operations.google_analytics_tag !== '', googleAnalyticsTag: config.operations.google_analytics_tag };
         }
     }catch( e ){
         logger.error(loadSettings.name, getCurrentLineNumber(), "Could not open/parse configuration file: " + process.argv[2] );